backend/requirements.txt
Django==5.2.9
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
python-dotenv==1.0.1
